+++ src/workspace/entities/workspace.entity.ts
@@ -103,6 +103,13 @@ export class DbWorkspace extends BaseEntity {
   })
   public payee_user_id: number | null;
 
+  @ApiModelProperty({
+    description: "Flag indicating workspace has exceeded upgrade threshold",
+    example: true,
+  })
+  @Column({ default: false, select: false, insert: false })
+  public exceeds_upgrade_limits: boolean;
+
   @ApiHideProperty()
   @OneToMany(() => DbWorkspaceMember, (workspaceMember) => workspaceMember.workspace, { onDelete: "CASCADE" })
   public members: DbWorkspaceMember[];